feat : 모든 쿼리문 respository 로 이동
* 다른 모듈 제외하고 모든 쿼리문 이동
