+++ server/src/playlist/playlist.repository.ts
@@ -10,10 +10,13 @@ import { DataSource, QueryRunner, Repository } from 'typeorm';
 @Injectable()
 export class PlaylistRepository {
   private playlistRepository: Repository<Playlist>;
+  private music_playlistRepository: Repository<Music_Playlist>;
   private readonly logger: Logger = new Logger('PlaylistEntity');
 
   constructor(private readonly dataSource: DataSource) {
     this.playlistRepository = this.dataSource.getRepository(Playlist);
+    this.music_playlistRepository =
+      this.dataSource.getRepository(Music_Playlist);
   }
 
   async isExistPlaylistOnUser(
@@ -104,6 +107,60 @@ export class PlaylistRepository {
 
       this.logger.error(
         `playlist.repository - deleteSinglePlaylist : SERVICE_ERROR`,
+      );
+      throw new CatchyException(
+        'SERVICE_ERROR',
+        HTTP_STATUS_CODE.BAD_REQUEST,
+        ERROR_CODE.SERVICE_ERROR,
+      );
+    } finally {
+      await queryRunner.release();
+    }
+  }
+
+  async deleteMusicInPlaylist(
+    playlistId: number,
+    musicId: string,
+  ): Promise<number> {
+    const queryRunner = this.dataSource.createQueryRunner();
+    await queryRunner.startTransaction();
+
+    try {
+      const target: Music_Playlist =
+        await this.music_playlistRepository.findOne({
+          where: {
+            music: { music_id: musicId },
+            playlist: { playlist_id: playlistId },
+          },
+        });
+
+      const deletedMusicPlaylistId: number = target.music_playlist_id;
+
+      if (target == undefined) {
+        this.logger.error(
+          `playlist.service - deleteMusicInPlaylist : NOT_ADDED_MUSIC`,
+        );
+        throw new CatchyException(
+          'NOT_ADDED_MUSIC',
+          HTTP_STATUS_CODE.BAD_REQUEST,
+          ERROR_CODE.NOT_ADDED_MUSIC,
+        );
+      }
+
+      await queryRunner.manager.remove(target);
+
+      await queryRunner.commitTransaction();
+
+      return deletedMusicPlaylistId;
+    } catch (error) {
+      await queryRunner.rollbackTransaction();
+
+      if (error instanceof CatchyException) {
+        throw error;
+      }
+
+      this.logger.error(
+        `playlist.service - deleteMusicInPlaylist : SERVICE_ERROR`,
       );
       throw new CatchyException(
         'SERVICE_ERROR',
